+++ Projet TDG/Sommet.h
@@ -63,7 +63,7 @@ class Sommet
         void CalculPcci(Sommet* si, Sommet* s_arrive, bool oriente);
         void Normaliser(int n);
         void AfficherTout();
-        void setSommet(int &indice, char &nom, int &x, int &y);
+        void setSommet(int &indice, char &nom, double &x, double &y);
         void SuppSommet(int &id2);
 
         void SaveComparaison(std::ofstream& fichier);
